Refactor template things into template.py
