+++ nachomemes/store.py
@@ -5,7 +5,6 @@ from pathlib import Path
 from sys import maxsize
 from typing import IO, Callable, Iterable, List, Optional, Union, Dict
 from urllib.request import Request, urlopen
-from tempfile import NamedTemporaryFile
 import atexit, os, re
 
 from discord import Member, Role, Guild
@@ -13,72 +12,7 @@ from dacite import Config
 from PIL import Image, ImageFont
 
 from .guild_config import GuildConfig
-from .style import Color, Justify, Font, TextBox
-
-# Monkeypatch Request to show the url in repr
-Request.__repr__ = lambda self: f"Request(<{self.full_url}>)"
-
-
-LOCAL_FILE_CACHE: Dict[str,str] = {}
-
-@atexit.register
-def _delete_cache():
-    for f in LOCAL_FILE_CACHE.values():
-        os.remove(f)
-
-
-def _fetch_image(url: Request) -> IO:
-    if url.type == 'file':
-        return urlopen(url)
-    if url.full_url not in LOCAL_FILE_CACHE:
-        print(url.full_url)
-        suffix=re.sub(r'[\W]', '', url.full_url.split('.')[-1])[:5]
-        with NamedTemporaryFile(suffix='.'+suffix, delete=False) as f:
-            with urlopen(url) as u:
-                f.write(u.read())
-            LOCAL_FILE_CACHE[url.full_url] = f.name
-        print(f.name)
-    return open(LOCAL_FILE_CACHE[url.full_url], 'rb')
-
-    
-
-
-class TemplateError(Exception):
-    pass
-
-@dataclass
-class MemeTemplate:
-    "Anatomy of a Meme"
-
-    name: str
-
-    # URL to load the image
-    source_image_url: Request
-
-    # Where to put the text
-    textboxes: List[TextBox]
-
-    # name of the layout
-    layout: str
-    description: str
-    docs: str
-
-    # times used
-    usage: int = 0
-
-    def read_source_image(self, buffer) -> Image:
-        with _fetch_image(self.source_image_url) as s:
-            buffer.write(s.read())
-            buffer.flush
-            buffer.seek(0)
-            return Image.open(buffer)
-
-    def render(self, message: Iterable[str], output: IO):
-        from nachomemes import render_template
-        render_template(self, message, output)
-
-
-
+from .template import Template, Font, Color, Justify
 
 # additional deserializers for dacite
 da_config = Config(
@@ -92,7 +26,6 @@ da_config = Config(
     }
 )
 
-
 class Store(ABC):
     @abstractmethod
     def refresh_memes(self, guild: Optional[Guild], hard: bool = False) -> str:
@@ -101,7 +34,7 @@ class Store(ABC):
     @abstractmethod
     def read_meme(
         self, guild: Optional[Guild], id: str, increment_use: bool = False
-    ) -> MemeTemplate:
+    ) -> Template:
         pass
 
     @abstractmethod
